emproving code quality

+++ app/Models/Event/EventVenue.php
@@ -11,9 +11,78 @@ use Illuminate\Database\Eloquent\SoftDeletes;
 use Spatie\Activitylog\LogOptions;
 use Spatie\Activitylog\Traits\LogsActivity;
 
+/**
+ * @property int $id
+ * @property string $uuid
+ * @property string $name
+ * @property string|null $description
+ * @property string $address_line1
+ * @property string|null $address_line2
+ * @property string $city
+ * @property string|null $state
+ * @property string|null $postal_code
+ * @property string $country
+ * @property numeric|null $latitude
+ * @property numeric|null $longitude
+ * @property int|null $total_capacity
+ * @property string|null $contact_name
+ * @property string|null $contact_email
+ * @property string|null $contact_phone
+ * @property string|null $website
+ * @property array<array-key, mixed>|null $amenities
+ * @property array<array-key, mixed>|null $images
+ * @property string|null $access_instructions
+ * @property string|null $parking_info
+ * @property bool $is_active
+ * @property \Illuminate\Support\Carbon|null $created_at
+ * @property \Illuminate\Support\Carbon|null $updated_at
+ * @property \Illuminate\Support\Carbon|null $deleted_at
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, \Spatie\Activitylog\Models\Activity> $activities
+ * @property-read int|null $activities_count
+ * @property-read int $available_rooms_count
+ * @property-read string $full_address
+ * @property-read int|null $rooms_count
+ * @property-read int $total_room_capacity
+ * @property-read \Illuminate\Database\Eloquent\Collection<int, \App\Models\Event\VenueRoom> $rooms
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue active()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue inCity(string $city)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue onlyTrashed()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue query()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereAccessInstructions($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereAddressLine1($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereAddressLine2($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereAmenities($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereCity($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereContactEmail($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereContactName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereContactPhone($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereCountry($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereCreatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereDeletedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereDescription($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereId($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereImages($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereIsActive($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereLatitude($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereLongitude($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereName($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereParkingInfo($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue wherePostalCode($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereState($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereTotalCapacity($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereUpdatedAt($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereUuid($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue whereWebsite($value)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue withCapacityAtLeast(int $capacity)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue withTrashed(bool $withTrashed = true)
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|EventVenue withoutTrashed()
+ * @mixin \Eloquent
+ */
 class EventVenue extends Model
 {
-    use HasFactory, HasUuid, LogsActivity, ClearsCache, SoftDeletes;
+    use ClearsCache, HasFactory, HasUuid, LogsActivity, SoftDeletes;
 
     protected $fillable = [
         'name',
@@ -89,7 +158,7 @@ class EventVenue extends Model
         $parts = [
             $this->address_line1,
             $this->address_line2,
-            $this->postal_code . ' ' . $this->city,
+            $this->postal_code.' '.$this->city,
             $this->country,
         ];
 
@@ -108,7 +177,7 @@ class EventVenue extends Model
 
     public function getTotalRoomCapacityAttribute(): int
     {
-        return $this->rooms()->sum('capacity') ?? 0;
+        return (int) ($this->rooms()->sum('capacity') ?? 0);
     }
 
     // Methods
@@ -124,7 +193,7 @@ class EventVenue extends Model
             return "https://www.google.com/maps?q={$this->latitude},{$this->longitude}";
         }
 
-        return "https://www.google.com/maps/search/?api=1&query=" . urlencode($this->full_address);
+        return 'https://www.google.com/maps/search/?api=1&query='.urlencode($this->full_address);
     }
 
     public function hasAmenity(string $amenity): bool
